test: ControlTest pour in + out
